Add Loger To Context

+++ SamayaElectronics.Api/DTO/PageDto.cs
@@ -10,11 +10,11 @@ namespace SamayaElectronics.Api.DTO
     public class PageDto<T>
     {
         public IEnumerable<T> Data { get; set; }
-        public int TotalCount { get; set; }
+        public float TotalCount { get; set; }
 
-        public int TotalPages { get; set; }
+        public float TotalPages { get; set; }
 
-        public int CurrentPage { get; set; }
+        public float CurrentPage { get; set; }
 
         public string PreviousLink { get; set; }
